backend/api-gateway/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic-settings>=2.11.0
python-jose[cryptography]==3.3.0
prometheus-client==0.23.1
